feat(resource): Implement ResourceVoter security

+++ src/Controller/ResourceController.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace App\Controller\Admin;
+namespace App\Controller;
 
 use App\Entity\Resource;
 use App\Form\ResourceType;
 use App\Repository\CategoryRepository;
 use App\Repository\ResourceRepository;
+use App\Security\Voter\ResourceVoter;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -15,7 +16,7 @@ use Symfony\Component\Routing\Annotation\Route;
 #[Route('/admin/resource')]
 class ResourceController extends AbstractController
 {
-    #[Route('/', name: 'admin_resource_index', methods: ['GET'])]
+    #[Route('/', name: 'app_resource_index', methods: ['GET'])]
     public function index(
         Request $request,
         ResourceRepository $resourceRepository,
@@ -31,16 +32,18 @@ class ResourceController extends AbstractController
             $resources = $resourceRepository->findBy([], ['createdAt' => 'DESC']);
         }
 
-        return $this->render('admin/resource/index.html.twig', [
+        return $this->render('resource/index.html.twig', [
             'resources' => $resources,
             'categories' => $categories,
             'selectedCategoryId' => $selectedCategoryId
         ]);
     }
 
-    #[Route('/new', name: 'admin_resource_new', methods: ['GET', 'POST'])]
+    #[Route('/new', name: 'app_resource_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted(ResourceVoter::CREATE);
+
         $resource = new Resource();
         $form = $this->createForm(ResourceType::class, $resource);
         $form->handleRequest($request);
@@ -55,18 +58,30 @@ class ResourceController extends AbstractController
 
             $this->addFlash('success', 'La ressource a été créée avec succès.');
 
-            return $this->redirectToRoute('admin_resource_index');
+            return $this->redirectToRoute('app_resource_index');
         }
 
-        return $this->render('admin/resource/new.html.twig', [
+        return $this->render('resource/new.html.twig', [
             'resource' => $resource,
             'form' => $form,
         ]);
     }
 
-    #[Route('/{id}/edit', name: 'admin_resource_edit', methods: ['GET', 'POST'])]
+    #[Route('/{id}', name: 'app_resource_show', methods: ['GET'])]
+    public function show(Resource $resource): Response
+    {
+        $this->denyAccessUnlessGranted(ResourceVoter::VIEW, $resource);
+
+        return $this->render('resource/show.html.twig', [
+            'resource' => $resource,
+        ]);
+    }
+
+    #[Route('/{id}/edit', name: 'app_resource_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Resource $resource, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted(ResourceVoter::EDIT, $resource);
+
         $form = $this->createForm(ResourceType::class, $resource);
         $form->handleRequest($request);
 
@@ -77,18 +92,20 @@ class ResourceController extends AbstractController
 
             $this->addFlash('success', 'La ressource a été modifiée avec succès.');
 
-            return $this->redirectToRoute('admin_resource_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('app_resource_index', [], Response::HTTP_SEE_OTHER);
         }
 
-        return $this->render('admin/resource/edit.html.twig', [
+        return $this->render('resource/edit.html.twig', [
             'resource' => $resource,
             'form' => $form,
         ]);
     }
 
-    #[Route('/{id}', name: 'admin_resource_delete', methods: ['POST'])]
+    #[Route('/{id}', name: 'app_resource_delete', methods: ['POST'])]
     public function delete(Request $request, Resource $resource, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted(ResourceVoter::DELETE, $resource);
+
         // Vérifie le jeton CSRF pour s'assurer que la requête provient bien de notre application
         if ($this->isCsrfTokenValid('delete'.$resource->getId(), $request->request->get('_token'))) {
             $entityManager->remove($resource); // Marque l'entité pour la suppression
@@ -98,6 +115,6 @@ class ResourceController extends AbstractController
             $this->addFlash('error', 'Jeton CSRF invalide. La suppression a été annulée');
         }
 
-        return $this->redirectToRoute('admin_resource_index', [], Response::HTTP_SEE_OTHER);
+        return $this->redirectToRoute('app_resource_index', [], Response::HTTP_SEE_OTHER);
     }
 }
